requirements.txt
absl-py==2.1.0
accelerate==0.29.2
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
astroid==3.2.2
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
blis==0.7.11
catalogue==2.0.10
catboost==1.2.7
certifi==2024.2.2
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
cloudpickle==3.0.0
colorama==0.4.6
confection==0.1.4
contourpy==1.2.1
cycler==0.12.1
cymem==2.0.8
dill==0.3.8
dnspython==2.6.1
docx2txt==0.8
email_validator==2.1.1
emoji==2.14.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
fake-useragent==1.5.1
fastapi==0.111.0
fastapi-cli==0.0.3
filelock==3.13.4
flatbuffers==24.12.23
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.3.1
gast==0.6.0
gensim==4.3.2
google-pasta==0.2.0
googletrans==4.0.0rc1
graphviz==0.20.3
grpcio==1.68.1
h11==0.9.0
h2==3.2.0
h5py==3.12.1
hiredis==2.3.2
hpack==3.0.0
hstspreload==2024.4.1
httpcore==0.9.1
httptools==0.6.1
httpx==0.13.3
huggingface-hub==0.27.0
hyperframe==5.2.0
idna==2.10
imbalanced-learn==0.13.0
imblearn==0.0
iniconfig==2.0.0
isort==5.13.2
Jinja2==3.1.3
joblib==1.4.0
keras==3.7.0
kiwisolver==1.4.5
langcodes==3.4.0
language_data==1.2.0
libclang==18.1.1
lightgbm==4.5.0
llvmlite==0.42.0
lxml==5.2.2
marisa-trie==1.1.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mccabe==0.7.0
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
multidict==6.0.5
murmurhash==1.0.10
namex==0.0.8
networkx==3.3
nltk==3.8.1
numba==0.59.1
numpy==1.26.4
opt_einsum==3.4.0
optree==0.13.1
orjson==3.10.3
packaging==24.0
pandas==2.2.2
pendulum==3.0.0
pillow==10.3.0
platformdirs==4.2.2
plotly==5.24.1
pluggy==1.5.0
preshed==3.0.9
protobuf==5.29.2
psutil==5.9.8
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
pyparsing==3.1.2
PyPDF2==3.0.1
pytest==8.2.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
redis==5.0.4
regex==2023.12.25
requests==2.31.0
rfc3986==1.5.0
rich==13.7.1
safetensors==0.4.2
scikit-learn==1.4.2
scipy==1.12.0
sentence-transformers==3.3.1
shap==0.45.1
shellingham==1.5.4
six==1.16.0
sklearn-compat==0.1.3
slicer==0.0.8
smart-open==6.4.0
sniffio==1.3.1
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
starlette==0.37.2
sympy==1.13.1
T0==3.1.5.6
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_intel==2.18.0
termcolor==2.5.0
textblob==0.18.0.post0
thinc==8.2.3
threadpoolctl==3.4.0
time-machine==2.14.1
tokenizers==0.21.0
tomlkit==0.12.5
torch==2.7.0.dev20250110+cu126
torchaudio==2.6.0.dev20250111+cu126
torchvision==0.22.0.dev20250111+cu126
tqdm==4.66.2
transformers==4.47.1
typer==0.12.3
typing_extensions==4.11.0
tzdata==2024.1
ujson==5.9.0
Unidecode==1.3.8
urllib3==2.2.1
uvicorn==0.29.0
wasabi==1.1.2
watchfiles==0.21.0
weasel==0.3.4
websockets==12.0
Werkzeug==3.1.3
wrapt==1.16.0
xgboost==2.1.3
yarl==1.9.4
